# Improve serialize dataTable performance